+++ src/index.ts
@@ -23,6 +23,22 @@ Module._extensions['.js'] = function (module, filename) {
 
 import {NodeSDK} from "./lib/NodeSDK";
 
+/*
+//let RainbowSdk : any = NodeSDK;
+//NodeSDK.RainbowSdk = NodeSDK;
 NodeSDK.NodeSDK = NodeSDK;
 module.exports = NodeSDK;
-export { NodeSDK as NodeSDK};
+
+//let RainbowSdk = NodeSDK;
+//module.exports = RainbowSdk;
+//module.exports.NodeSDK = NodeSDK;
+
+export default NodeSDK;
+// */
+
+module.exports = NodeSDK;
+module.exports.default = NodeSDK;
+module.exports.NodeSDK = NodeSDK;
+export default NodeSDK;
+export { NodeSDK as NodeSDK}; //, OptionsType};
+