+++ CHANGELOG.md
@@ -18,7 +18,8 @@ All notable changes to Rainbow-Node-SDK will be documented in this file.
 -   Add `ContactsService::getListPersonalDirectoryEntriesData` This API allows connected users to get an entry of his personal directory. 
 -   Add `ContactsService::updatePersonalDirectoryEntry` This API allows the connected user to update an entry of his personnal directory. 
 -   Add `ContactsService::deletePersonalDirectoryEntry` This API allows connected user to delete an entry from his personal directory.
--   Fix logs when a restart of the SDK is done. 
+-   Fix logs when a restart of the SDK is done.
+-   Fix default typescript export. 
 
 ## [2.8.2] - 2022-02-18
 -   Add documentation about telephony call : `Telephony_calls.md`.